Added lots of features
Removed all of the dummy content. Added search feature with alerts, pagination, and a global object to track the current page etc.

API keys were moved to global object. Obviously this is just a demo, so that's not a big deal with a Free Public API.

Might add more to this later.